:label: update ui home and detail - from dollar to idr

+++ core/src/main/kotlin/com/rivaldy/id/core/data/model/Product.kt
@@ -17,7 +17,7 @@ data class Product(
     @SerializedName("images")
     val images: List<String?>? = null,
     @SerializedName("price")
-    val price: Int? = null,
+    val price: Long? = null,
     @SerializedName("rating")
     val rating: Double? = null,
     @SerializedName("stock")
